+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @nexus-state/core
 
+## 0.1.5
+
+### Patch Changes
+
+- add devtools functionality and performance
+
 ## 0.1.4
 
 ### Patch Changes