+++ packages/devtools/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @nexus-state/devtools
 
+## 0.1.4
+
+### Patch Changes
+
+- add devtools functionality and performance
+- Updated dependencies
+  - @nexus-state/core@0.1.5
+
 ## 0.1.3
 
 ### Patch Changes